Wrap handlers in transaction scope is now a separate feature
Stop using the funky ContinerRegistration

Approve changes
