+++ src/NServiceBus.Core/ConsistencyGuarantees/TransactionScopes/SuppressAmbientTransactionBehavior.cs
