+++ src/NServiceBus.Core/ConsistencyGuarantees/TransactionScopes/WrapHandlersInTransactionScope.cs
@@ -1,0 +1,33 @@
+﻿namespace NServiceBus.Features
+{
+    using System;
+    using System.Transactions;
+
+    class WrapHandlersInTransactionScope : Feature
+    {
+        public WrapHandlersInTransactionScope()
+        {
+            EnableByDefault();
+        }
+
+        protected internal override void Setup(FeatureConfigurationContext context)
+        {
+            if (context.Settings.GetOrDefault<bool>("Transactions.DoNotWrapHandlersExecutionInATransactionScope"))
+            {
+                context.Pipeline.Register<SuppressAmbientTransactionBehavior.Registration>();
+            }
+            else
+            {
+                var transactionOptions = new TransactionOptions
+                {
+                    IsolationLevel = context.Settings.Get<IsolationLevel>("Transactions.IsolationLevel"),
+                    Timeout = context.Settings.Get<TimeSpan>("Transactions.DefaultTimeout")
+                };
+
+                context.Container.ConfigureComponent(b => new HandlerTransactionScopeWrapperBehavior(transactionOptions), DependencyLifecycle.InstancePerCall);
+
+                context.Pipeline.Register<HandlerTransactionScopeWrapperBehavior.Registration>();
+            }
+        }
+    }
+}
