+++ src/NServiceBus.Core/Unicast/Transport/TransactionSettings.cs
@@ -7,7 +7,7 @@
     /// <summary>
     /// Settings related to transactions.
     /// </summary>
-    public class TransactionSettings
+    public partial class TransactionSettings
     {
         internal TransactionSettings(ReadOnlySettings settings)
         {
@@ -15,28 +15,8 @@
             TransactionTimeout = settings.Get<TimeSpan>("Transactions.DefaultTimeout");
             IsolationLevel = settings.Get<IsolationLevel>("Transactions.IsolationLevel");
             SuppressDistributedTransactions = settings.Get<bool>("Transactions.SuppressDistributedTransactions");
-            DoNotWrapHandlersExecutionInATransactionScope = settings.Get<bool>("Transactions.DoNotWrapHandlersExecutionInATransactionScope");
         }
 
-        /// <summary>
-        /// Create a new settings.
-        /// </summary>
-        /// <param name="isTransactional">Is transactions on.</param>
-        /// <param name="transactionTimeout">The tx timeout.</param>
-        /// <param name="isolationLevel">The isolation level.</param>
-        /// <param name="suppressDistributedTransactions">Should DTC be suppressed.</param>
-        /// <param name="doNotWrapHandlersExecutionInATransactionScope">Should handlers be wrapped.</param>
-        public TransactionSettings(bool isTransactional, TimeSpan transactionTimeout, IsolationLevel isolationLevel, bool suppressDistributedTransactions, bool doNotWrapHandlersExecutionInATransactionScope)
-        {
-            Guard.AgainstNegative("transactionTimeout", transactionTimeout);
-            IsTransactional = isTransactional;
-            TransactionTimeout = transactionTimeout;
-            IsolationLevel = isolationLevel;
-            SuppressDistributedTransactions = suppressDistributedTransactions;
-            DoNotWrapHandlersExecutionInATransactionScope = doNotWrapHandlersExecutionInATransactionScope;
-        }
-
-     
         /// <summary>
         /// Sets whether or not the transport is transactional.
         /// </summary>
@@ -58,10 +38,5 @@
         /// If true the transport won't enlist in distributed transactions.
         /// </summary>
         public bool SuppressDistributedTransactions { get; set; }
-
-        /// <summary>
-        /// Controls if the message handlers should be wrapped in a <see cref="TransactionScope"/>.
-        /// </summary>
-        public bool DoNotWrapHandlersExecutionInATransactionScope { get; set; }
     }
 }
